### 2025-02-25 #7847 @editorialbot
[PRE REVIEW]: BlueCarbon R package: Estimation of Organic Carbon Stocks and Sequestration Rates From Soil Core Data
**Submitting author:** @NPJuncal (<a href="http://orcid.org/0000-0003-3767-1812">Nerea Piñeiro-Juncal</a>)
**Repository:** https://github.com/EcologyR/BlueCarbon
**Branch with paper.md** (empty if default branch): master
**Version:** v1.0
