### 2025-03-06 ed4a84e
add logo. Close #48
