+++ README.Rmd
@@ -16,7 +16,7 @@ knitr::opts_chunk$set(
 )
 ```
 
-# BlueCarbon
+# BlueCarbon <a href="https://ecologyr.github.io/BlueCarbon/"><img src="man/figures/logo.png" align="right" height="126" alt="BlueCarbon website" /></a>
 
 <!-- badges: start -->
 
