+++ pkgdown.yml
@@ -3,7 +3,7 @@ pkgdown: 2.1.1
 pkgdown_sha: ~
 articles:
   articles/Tutorial: Tutorial.html
-last_built: 2025-02-24T15:24Z
+last_built: 2025-03-06T12:14Z
 urls:
   reference: https://ecologyr.github.io/BlueCarbon/reference
   article: https://ecologyr.github.io/BlueCarbon/articles
