### 2025-04-23 #7847 @goldingn
Hi @Qepanna, would you be interested in reviewing this R package submission for JOSS? It's a set of tools for estimating carbon stocks from soil core data.
