### 2025-04-23 #7847 @goldingn
Hi @Tania-Maxwell, would you be interested in reviewing this R package submission for estimating carbon stocks from soil core data, for the Journal of Open Source Software?
