### 2025-04-23 #7847 @NPJuncal
> Hi [@NPJuncal](https://github.com/NPJuncal), could you please suggest some potential reviewers? Either people you know of (who are on GitHub) or from searching [the list of JOSS reviewers](https://reviewers.joss.theoj.org/lookup). Please mention their GitHub handles here in this thread (without tagging them with an @). You can search that have already agreed to review and may be suitable for this submission.

Hi @goldingn , maybe kanishkan91 or jordipages-repo could be good reviewers.
