### 2025-04-23 #7847 @goldingn
Hi @kanishkan91 would you be interested in reviewing this this R package submission for estimating carbon stocks from soil core data, for the Journal of Open Source Software?
