### 2025-04-24 #7847 @goldingn
Hi @jordipages-repo, that's great, thanks for your interest!

It's not a problem that you haven't written an R package before. As an experienced R programmer with domain expertise, you'll be able to review usefulness of the package. We'll make sure that at least one other reviewer has experience writing R packages.

The review process is very straightforward and well guided. Once we have another reviewer lined up, I'll kick off the review process in a new thread. You'll have a checklist to work through, [like this one](https://github.com/openjournals/joss-reviews/issues/7852#issuecomment-2684149970) and we'll all be in the same thread to discuss issues and opportunities for improvement together. Specific changes can be discusses ad issues on the project GitHub page. There's also a manual for reviewers [here](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html)

We'd hope you could finalise your review within 6 weeks, though the sooner you can start the review and start flagging issues, the quicker the authors can resolve them and the more enjoyable the process is for everyone.

Happy to answer any more questions
