### 2025-04-25 #7847 @Tania-Maxwell
Hi @goldingn , thank you for considering me! Looks like a very useful package. I'm just back from maternity leave and I'm very busy catching up with things, but I would be happy to be considered for future reviews for the Journal of Open Source Software.
