### 2025-04-28 #7847 @goldingn
Hi @doserjef, would you be interested in reviewing this R package submission for estimating carbon stocks from soil core data, for the Journal of Open Source Software?
