### 2025-04-28 #7847 @doserjef
Hi @goldingn, thanks for the note. I unfortunately do not have the time at the moment to add on another review, but please do keep me in mind for future reviews at JOSS.
