### 2025-07-04 #8333 @jordipages-repo
Hi @goldingn,

I've just started my checklist, and I’d like to flag a possible conflict of interest with the lead author. We used to work at the same  institution (until 2019 or 2020), and we are currently both authors on a review paper where I’ve contributed some data. Given this context, I’d appreciate your guidance on whether it’s appropriate for me to continue as a reviewer.

Thanks!
