### 2025-07-11 #8333 @goldingn
Thank you both for clarifying. Even though the paper has many authors, since @NPJuncal is the lead author of the paper, I do think that's a close collaboration and a conflict of interest.

The next steps from here:
- I'll remove @jordipages-repo as a reviewer
- I'll search for a replacement reviewer
- @smwindecker, in the mean time please continue with your review

It was a little tricky to find a reviewer earlier, but there was at least one who was unavailable but may now be available. If we don't have any luck in the next week of so, I'd appreciate any help you can provide in suggesting more reviewers @NPJuncal.

And thanks again @jordipages-repo for agreeing to review in the first place and for letting me know about the COI!
