### 2025-07-11 EcologyR/BlueCarbon#50 @smwindecker
vignette doesn't build
Hi authors, this may be intentional, but the vignette is currently ignored in the RBuildIgnore file, and therefore can't be easily accessed by a user that installs the package and tries to find it using `browseVignettes('BlueCarbon')` or `vignette(package = 'BlueCarbon')` (both return "no vignettes found"). I can see it is on the website but if there's a way to have it build the traditional way that would be helpful as well.

https://github.com/openjournals/joss-reviews/issues/8333
