### 2025-07-11 EcologyR/BlueCarbon#51 @smwindecker
Blue Carbon package workflow image not rendering well in dark mode
Hi authors, please note the first image (Blue Carbon package workflow) is not rendering well in dark mode. The black lines and text are becoming invisible. Perhaps these can be given another color or a white background to avoid the problem and ensure users can read the flow.

https://github.com/openjournals/joss-reviews/issues/8333
