### 2025-07-11 #8333 @kthyng
@goldingn Sorry for my delay, but I agree with your ultimate assessment.
