### 2025-07-14 51e49c8
White background for workflow image #51
