+++ vignettes/articles/Tutorial.Rmd
@@ -59,9 +59,14 @@ head(compaction, n = 10)
 
 ## Decompact cores
 
-Core compaction does not affect the final stock of the core. However, it does affect the depth of the core samples. As stocks are given until a known soil depth (e.g. the first meter) to be able to compare among areas, the depth of the samples is needed to properly estimate the final stock.
+Core compaction does not affect the final stock of the core.
+However, it does affect the depth of the core samples.
+As stocks are given until a known soil depth (e.g. the first meter) to be able to compare among areas, the depth of the samples is needed to properly estimate the final stock.
 
-We use function `decompact()` to estimate the decompacted minimum and maximum depth of each sample before stock estimation. Furthermore, as compaction reduces the volume of the samples, dry bulk density has to be corrected. This can be done by estimating the corrected volume by: multiplying the corrected thickness of the sample (corrected maximum depth - corrected minimum depth) by the area of the core; or, if compacted dry bulk density is provided, function `decompact()` will correct it. To provide compacted dry bulk density the name of the column has to be specified in the function parameter "dbd" as, by default, this parameter is NULL and the function assumes that it was not provided.
+We use function `decompact()` to estimate the decompacted minimum and maximum depth of each sample before stock estimation.
+Furthermore, as compaction reduces the volume of the samples, dry bulk density has to be corrected.
+This can be done by estimating the corrected volume by: multiplying the corrected thickness of the sample (corrected maximum depth - corrected minimum depth) by the area of the core; or, if compacted dry bulk density is provided, function `decompact()` will correct it.
+To provide compacted dry bulk density the name of the column has to be specified in the function parameter "dbd" as, by default, this parameter is NULL and the function assumes that it was not provided.
 
 `decompact()` has 6 parameters:
 
@@ -77,13 +82,15 @@ We use function `decompact()` to estimate the decompacted minimum and maximum de
 
 -   "dbd": (character) Name of the column with dry bulk density
 
-Here we don't need to specify the name of the columns as in the example data the names are the same as the default names in the function. But we DO need to specify the dry bulk density column name as by default this argument is NULL.
+Here we don't need to specify the name of the columns as in the example data the names are the same as the default names in the function.
+But we DO need to specify the dry bulk density column name as by default this argument is NULL.
 
 ```{r}
 bluecarbon_decompact <- decompact(bluecarbon_data, dbd = "dbd")
 ```
 
-The warning has given a list of cores that have no compaction percentage data. The function assumes that these cores were not compacted in the field and sets compaction to zero.
+The warning has given a list of cores that have no compaction percentage data.
+The function assumes that these cores were not compacted in the field and sets compaction to zero.
 
 Here are the first 10 rows of the final dataframe, with corrected minimum and maximum sample depth and dry bulk density.
 
@@ -92,17 +99,25 @@ head(bluecarbon_decompact, n = 10) |>
   kable()
 ```
 
-
-
 ## Estimation of organic carbon from organic matter
 
-There is a well-known linear relation between organic matter content and organic carbon in soils. This correlation can change between ecosystems and sampling sites due to changes in organic matter composition among other factors. Therefore, it is recommended to estimate this correlation for each batch of samples.
+There is a well-known linear relation between organic matter content and organic carbon in soils.
+This correlation can change between ecosystems and sampling sites due to changes in organic matter composition among other factors.
+Therefore, it is recommended to estimate this correlation for each batch of samples.
 
[…]
