+++ pkgdown.yml
@@ -1,9 +1,9 @@
 pandoc: 3.1.11
-pkgdown: 2.1.1
+pkgdown: 2.1.3
 pkgdown_sha: ~
 articles:
   articles/Tutorial: Tutorial.html
-last_built: 2025-03-06T12:14Z
+last_built: 2025-07-14T14:13Z
 urls:
   reference: https://ecologyr.github.io/BlueCarbon/reference
   article: https://ecologyr.github.io/BlueCarbon/articles
