### 2025-07-14 e535b8b
updated references
