+++ manuscript/paper.bib
@@ -15,11 +15,12 @@
 	file = {PDF:C\:\\Users\\npjun\\Zotero\\storage\\3DXTZIZL\\1-s2.0-S0964569123000200-main.pdf:application/pdf}
 }
 
-@article{Howard_2014,
-	title = {Coastal Blue Carbon: Methods for assessing carbon stocks and emissions factors in mangroves, tidal salt marshes, and seagrass meadows.},
-	author = {Howard, J. and Hoyt, S. and Isensee, K. and Pidgeon, E. and Telszewski, M.},
-	journal = { Conservation International, Intergovernmental Oceanographic Commission of UNESCO, International Union for Conservation of Nature},
-	year = {2014}
+@book{Howard_2014,
+  	url = {https://www.thebluecarboninitiative.org/manual},
+  	Author = {Howard, J. and Hoyt, S. and Isensee, K. and Pidgeon, E. and Telszewski, M.},
+  	Booktitle = {Coastal Blue Carbon: Methods for assessing carbon stocks and emissions factors in mangroves, tidal salt marshes, and seagrass meadows.},
+  	Publisher = {Conservation International, Intergovernmental Oceanographic Commission of UNESCO, International Union for Conservation of Nature},
+  	Year = 2014
 }
 
 @article{Fourqurean_2012,
@@ -56,12 +57,16 @@ ISBN: 1752-0894},
 
 
 @article{Piñeiro-Juncal,
-	title = {Soil organic carbon preservation and decay trends in tidal
-marsh, mangrove and seagrass blue carbon ecosystems},
+	title = {Soil organic carbon depth profiles and centennial and millennial decay rates in tidal marsh, mangrove and seagrass blue carbon ecosystems},
+	volume = {6},
+	issn = {2662-4435},
+	doi = {10.1038/s43247-025-02390-2},
+	abstract = {The management of coastal blue carbon ecosystems can contribute to mitigate anthropogenic greenhouse gas emissions. However, limited information on soil organic carbon (OC) decay rates in tidal marsh, mangrove, and seagrass soils hinders their inclusion in climate strategies and carbon-crediting schemes. Here, we analyzed downcore OC trends in 3761 soil cores from blue carbon ecosystems worldwide. A decrease in OC content with soil depth was measured in 63% of the cores, whereas stable and increasing trends were observed in 22% and 15% of the cores, respectively. Based on 85 profiles where OC decay could be modelled, the OC decay rate in blue carbon ecosystems was 0.023 ± 0.002 yr−1 over the last 100 years and 0.007 ± 0.0007 yr−1 over the last 1000 years. This results in the stabilization of 9% and 0.1% of the soil OC inputs 100 and 1000 yr after burial, respectively, showcasing the long residence time of OC in the sinks associated to blue carbon ecosystems. This study provides baseline estimates of soil OC decay under natural conditions for short- and long-term periods, and to assess the carbon abatement and additionally provided by carbon offsetting projects in blue carbon ecosystems.},
+	number = {1},
+	journal = {Communications Earth & Environment},
 	author = {Piñeiro-Juncal, Nerea and Mateo, Miguel Ángel and Leiva-Dueñas, Carmen and Serrano, Eduard and Inostroza, Karina and Soler, Montserrat and Apostolaki, Eugenia T. and Lavery, Paul and Duarte, Carlos M. and Lafratta, Anna and Serrano, Oscar},
-	year = {under review}
+	year = {2025}
 }
-
 
 
 
