### 2025-07-15 EcologyR/BlueCarbon#52 @smwindecker
goodpractice checks outcome
Hi authors, some checks I ran returned these suggestions. They are for your consideration only -

   ✖ avoid sapply(), it is not type safe. It might return a vector, or a list, depending on the input data.
    Consider using vapply() instead.

    R/test_extrapolation.R:111:12

  ✖ avoid 1:length(...), 1:nrow(...), 1:ncol(...), 1:NROW(...) and 1:NCOL(...) expressions. They are error
    prone and result 1:0 if the expression on the right hand side is zero. Use seq_len() or seq_along()
    instead.

    R/estimate_oc.R:113:26

https://github.com/openjournals/joss-reviews/issues/8333
