### 2025-07-15 1d0a758
Revert "edit typo in documentation"
This reverts commit 23eafe131f12be939460b36e2dea0023cab1f91f.
