### 2025-07-15 EcologyR/BlueCarbon#54 @smwindecker
citations that would be good
Hi authors, from the paper you've argued that these functions form a sort of tool box to handle a number of fairly run-of-the-mill methods implementations for doing these blue carbon core processing steps. I think given this package will sit in that methods context, some more citations to these well-known methods will help a user link these implementations to their basis in the blue carbon literature. Given this, I think a few extra methods citations would be good to add. I've listed some places below to add citations, if possible:

- linear correction for compaction: in the readme you refer to the compaction correction being linear. I think it would be worth adding a citation for the appropriateness of assuming linear impacts of compaction down the soil core, if available.
- linear correlation between organic carbon and organic matter content (line 104 of tutorial from what I can tell)
- linear extrapolation for accumulated mass of organic carbon and depth (in a separate issue I raise the option of allowing a user to substitute in a different functions, like a nonlinear one, for this relationship)

https://github.com/openjournals/joss-reviews/issues/8333
