### 2025-07-15 EcologyR/BlueCarbon#55 @smwindecker
stock sequestration tutorial instruction in need of rewrite
Line 286 of the tutorial explaining the sequestration function is confusing, it states: `The information of the model is included as the age of the sample...`

I think this means that the age of the soil sample is a required column to estimate sequestration rates. Is that correct? The sentence should be edited as it's worded in a confusing manner.

https://github.com/openjournals/joss-reviews/issues/8333
