### 2025-07-15 EcologyR/BlueCarbon#56 @smwindecker
estimate_oc_stock questions
1. extrapolating oc linearly down to _any_ depth:

from what I understand, the `test_extrapolation` function requires that some of the provided cores reach the desired standardisation depth, but I did not notice a similar maximum restriction for `estimate_oc_stock`. Extrapolating down linearly with depth seems risky with no bounds. Is it bound by at least one sample’s max depth? This line suggests it is not: "... estimate_oc_stock() as that function models those cores that do not reach the desired depth." If the authors would like to retain the option for an open-ended max depth for extrapolation perhaps they could provide a citation or two regarding reasonable max depth settings for different ecosystems, or something to that effect. Alternatively, the description could more directly point the users to checking with `test_extrapolation` if the extrapolation introduced excessive errors and amending their max depth choice accordingly.

2. extrapolating oc with depth _linearly_:

Have the authors considered making their `estimate_h` function adaptable so a user could introduce their own function with depth? In previous work I assisted colleagues to build a nonlinear model of OC with depth. Even if providing a suite of different functions is out of scope, it could be a really nice functionality to allow users to swap in a different function at this stage.

https://github.com/openjournals/joss-reviews/issues/8333
