### 2025-07-15 EcologyR/BlueCarbon#57 @smwindecker
organic carbon % terminology throughout tutorial
The first time I read through the tutorial I found the terminology for organic carbon at each step a bit confusing. Namely, we seem to be estimating organic carbon at the `estimate_oc` stage, but then we are using the following equation OC=DBD*(OCp/100)*h later on in the estimate stock stage which appears to also be estimating organic carbon. From what I could tell from closer inspection, the first phase is outputting % organic carbon, which is later used as OCp in the equation. I think this needs to be made more consistent in the language throughout the tutorial and documentation, for example:
  -  by changing the subheading "Estimation of organic carbon from organic matter" to "Estimate % organic carbon from % organic matter" (or something like that)
  - including the language of "OCp" before it is introduced in the equation
  - in line 174 "where DBD is the dry bulk density, OCp the organic carbon content in percentage and h the thickness of the sample" you could change to reference "OCp the organic carbon content in percentage as calculated in step X"...
  - in the description for the functions estimate_oc() and estimate_oc_stock(), it is unclear what units the 'om' and 'oc' columns should take
  - line 152 describes the outputs of estimate_oc: "three new columns on the estimated organic carbon, the standard error of the estimate..." but it's still unclear what the units are. Are these already % values? Needs to be stated and perhaps the output col name changed to OCp.
- line 111 “If an organic carbon value is already available for a sample, the function returns it.” Again this needs to be clear to the user that provided oc col should be % so that if it is used down the line it is the right units.
- etc

https://github.com/openjournals/joss-reviews/issues/8333
