### 2025-07-15 EcologyR/BlueCarbon#58 @smwindecker
DBD clarification
Adjusting DBD for compaction appears to be optional (the col default is set to NULL). But we later need to use a column for dbd to calculate stock. Is the assumption here that some folks would use the `decompact` functions without wanting to estimate stock (and therefore that's why they don't have a dbd col) or that folks would be electing to not account for compaction in the dbd column? I found this unclear and should be clarified as it seems like it gives the option for a user to supply compaction corrected oc but non corrected dbd.

https://github.com/openjournals/joss-reviews/issues/8333
