### 2025-07-15 EcologyR/BlueCarbon#59 @smwindecker
anticipating different user data
The tutorial currently reads quite well, and seems very useful to step through the steps needed in a use case very close to that which is described. But I can imagine some use cases where users would have a different sort of initial data. For example, some folks will have OC itself from getting soil samples directly measured, do they need to use the estimate_oc step or can they skip it? They may want to go straight to stock estimates. Some would want to go first to account for compaction then straight to stock, some others would have OM and want to use the functions to estimate OCp first then go to stock.

I think the tutorial could really benefit from two or three short sentences highlighting this non linearity given, from what I can see, the functions don't need to be run in sequence since the column names can be assigned at each step. I think this may also be helped by addressing https://github.com/EcologyR/BlueCarbon/issues/57 which will address some clarity on the units/data type for om and oc columns.

https://github.com/openjournals/joss-reviews/issues/8333
