+++ pkgdown.yml
@@ -3,7 +3,7 @@ pkgdown: 2.1.3
 pkgdown_sha: ~
 articles:
   articles/Tutorial: Tutorial.html
-last_built: 2025-07-14T14:13Z
+last_built: 2025-07-15T08:17Z
 urls:
   reference: https://ecologyr.github.io/BlueCarbon/reference
   article: https://ecologyr.github.io/BlueCarbon/articles
