### 2025-07-15 3815cf2
new workflow image with white background
