### 2025-07-16 34dafb5
added a few lines in the tutorial to highlight that the fnctions can be used individually #59
