### 2025-07-16 EcologyR/BlueCarbon#58 @NPJuncal
Some reasearchers estimate dbd from the weight of the samples and the compacted volumen. In this case they need to correct the dbd. But most researchers estimate the corrected depth of the samples before estimate the volume, and, therefore, the dbd  do not need to be corrected.
We added the option to correct dbd for those minority of researchers (to my knowledge they are a minority) that prefer to estimate dbd from uncorrected volumes and then correct dbd.
do you think it is necesary to explain this, @smwindecker ? maybe in the readme?
