### 2025-07-16 fa881ed
#57
Subheadings for estimate_oc() function to "Estimation of organic carbon % from organic matter %"

OCp to OC%

Piñeiro-Juncal 2025 citation has been updated in the tutorial, estimate_oc() function and README

in estimate_oc() the following has been added at the end: The OC values provided will have the same units as those provided initially for the OC and OM columns. We recommend to work in percentage, as the estimate_oc_stock() require OC values in percentage.
