### 2025-07-16 cd630db
#55 changed to : The age of each sample is included in a column of the input dataframe (column "age" in the example dataframe).
