### 2025-07-16 ec11a5c
#54 Howard 2014 reference added
