### 2025-07-16 f2ab947
#56 added a warning about the prediction of stocks with no maximun depth bounduaries
