### 2025-07-18 EcologyR/BlueCarbon#58 @smwindecker
Thanks for the clarification. That makes sense. I think it's clear why you've done it this way, but yes maybe one or two sentences to point the reader on this. It may be especially useful given that some of those looking for these helper functions may themselves be slightly less familiar with the procedure.
