### 2025-07-24 e03a953
avoid sapply and 1:nrow #52
