### 2025-07-24 EcologyR/BlueCarbon#52 @Pakillo
Thanks, good catch!

Fixed now
