### 2025-07-24 7654029
Update CITATION.cff
