### 2025-07-27 #8333 @goldingn
Hi @Tania-Maxwell and @kanishkan91, you had both previously indicated you might be open to reviewing this JOSS submission but were too busy at the time. We are in need of a second reviewer again - are either of you still available?
