### 2025-08-01 #8333 @goldingn
Hi @doserjef, we're again looking for a reviewer for this R package JOSS submission on estimating carbon stocks from soil core data. Is there any chance you now have time to review?
