### 2025-08-08 #8333 @goldingn
Hi @cdivittorio would you be interested in reviewing this software on estimating carbon stocks from soil core data for the Journal of Open Source Software?
