### 2025-08-27 #8333 @cdivittorio
> Hi [@cdivittorio](https://github.com/cdivittorio) would you be interested in reviewing this software on estimating carbon stocks from soil core data for the Journal of Open Source Software?

Thank you for the offer, but I have not worked with raw core data and am not qualified to complete this review.
