### 2025-09-10 EcologyR/BlueCarbon#60 @Tania-Maxwell
Minor edits tutorial
I'd recommend adding a comment on units at the beginning of the tutorial.rmd document, in addition to the comment in lines 230-231.

A few suggested corrections:

code line 56 "This are" to "Below are "
68: add "the" before function

110: add "function" after "estimate_oc"

151: the legend in the graph still shows Pineiro-Juncal et al. in prep but it might be the 2025 paper?

247: slightly change to make the sentence clearer: "model the stock using observed data within 90, 75, 50 and 25% of the total observed depth"

https://github.com/openjournals/joss-reviews/issues/8333
