### 2025-09-10 EcologyR/BlueCarbon#61 @Tania-Maxwell
Figures not rendering well in dark mode
Similar to the issue about the package workflow not rendering well in dark mode (https://github.com/EcologyR/BlueCarbon/issues/51), the other three figures in the read.me file also do not render well in dark model and would need a white background.

https://github.com/openjournals/joss-reviews/issues/8333
