### 2025-09-15 79be7e3
Clarification about dry bulk density correction added #58
