+++ manuscript/paper.md
@@ -84,7 +84,7 @@ Sampling soil cores by manual percussion often results in the compaction of the 
 
 This function applies a linear correction (assuming uniform compaction of the core material) to adjust the sample depth accurately. If dry bulk density data is provided, the function also corrects it accordingly.
 
-#### ***estimate_oc*** **- Organic carbon content estimation from organic carbon data**
+#### ***estimate_oc*** **- Organic carbon % estimation from organic matter %**
 
 There is a linear correlation between organic carbon and organic matter content. This correlation can vary across ecosystems and sampling sites. This function fits a linear regression model between organic matter and organic carbon content of the samples and predicts organic carbon values for samples where the latter information is missing. Estimation of organic carbon is performed using a linear regression between the logarithm of the organic carbon content and the logarithm of the organic matter content (log(organic carbon) \~ log(organic matter)), providing an organic carbon value for each organic matter value. It fits a model for each sampling station, dominant species, and ecosystem. If an organic carbon value is already available for a sample, the function returns it. Otherwise, it applies the model for the corresponding sampling station. If a model cannot be fitted for that station (e.g. because of limited sample size) or if the model fit is poor, the function instead applies the model for the dominant species. If no suitable species-level model exists, it then applies the ecosystem-level model. If no models are available at any of these levels, the function defaults to published models: @Fourqurean_2012 for seagrasses, @Maxwell_2023 for salt marshes, and @Piñeiro-Juncal for mangroves. It is unlikely, but possible, that the model predicts higher organic carbon than organic matter content. If this occurs, the function issues a warning, and it is recommended to discard that model.
 
