### 2025-09-15 50dd45e
clarifications about units
