### 2025-09-15 EcologyR/BlueCarbon#60 @NPJuncal
I have added the changes sugested by the reviewer.
151: the legend in the plot generated by function estimate_oc() has been updated to Piñeiro-Juncal et al. 2025 in the GitHub repository code. However, it wont be updated in CRAN until we upload the next version.
