### 2025-09-16 99b2659
white background to render in black mode #61
