+++ vignettes/articles/Tutorial.Rmd
@@ -34,7 +34,7 @@ The core field information and laboratory data includes: sampling site, core ID,
 
 Many field methods to extract soil cores can lead to the compaction of the material retrieved (e.g. manual percussion).
 
-![](images/compaction-01.png){width="497"}
+![](images/compaction-02.png){width="462"}
 
 The compaction percentage can be estimated knowing the difference between the original surface level of the soil and the surface level of the soil within the sampler after core insertion and before retrieval.
 All distances in the example data are provided in cm.
@@ -103,7 +103,7 @@ Here are the first 10 rows of the final dataframe, with corrected minimum and ma
 
 ```{r}
 head(bluecarbon_decompact, n = 10) |> 
-  kable()
+  table()
 ```
 
 ## Estimation of organic carbon % from organic matter %
@@ -191,7 +191,7 @@ The function `estimate_h()` distributes these empty spaces between the adjacent 
 This allows to account for both different sample thickness and different spacing between samples.
 Then, the organic carbon mass of that sample will be estimated following equation 1 with the corrected sample thickness.
 
-![Distribution of empty spaces between samples by estimate_h() function](images/estimate_h().png){width="401"}
+![Distribution of empty spaces between samples by estimate_h() function](images/estimate_h()-01.png){width="431"}
 
 The function `estimate_h()` is already incorporated within functions `estimate_oc_stock()`, `test_extrapolation()` and `estimate_seq_rate()`, and there is no need to run it beforehand.
 Sample depths in the example dataset is provided in cm.
@@ -251,8 +251,7 @@ head(stocks75, n = 10)
 
 With this function (`test_extrapolation()`) we can visualize the amount of error that may be introduced by extrapolating carbon stocks of short cores to a deeper standardization depth.
 This function uses those cores that reach the desired depth and tests what would be the error in the estimated carbon stock if we were to model it from the 90, 75, 50 and 25% of that depth.
-This function subsets the cores that reach the desired depth, estimates the complete stock (observed stock) to indicated desired depth, model the stock using observed data within 90, 75, 50 and 25% of the total observed depth
-estimates the stock from the linear relation of organic carbon accumulated mass and depth using the 90, 75, 50 and 25% length of the indicated desired depth, and compares the observed stock with the stocks estimated by extrapolation.
+This function subsets the cores that reach the desired depth, estimates the complete stock (observed stock) to indicated desired depth, model the stock using observed data within 90, 75, 50 and 25% of the total observed depth estimates the stock from the linear relation of organic carbon accumulated mass and depth using the 90, 75, 50 and 25% length of the indicated desired depth, and compares the observed stock with the stocks estimated by extrapolation.
 IT DOES NOT provide the error for those cores modeled by the function `estimate_oc_stock()` as that function models those cores that do not reach the desired depth.
 This function requires that some of your cores do reach the desired depth.
 
