+++ README.md
@@ -1,162 +1,96 @@
-
 <!-- README.md is generated from README.Rmd. Please edit that file -->
 
 # BlueCarbon <a href="https://ecologyr.github.io/BlueCarbon/"><img src="man/figures/logo.png" alt="BlueCarbon website" align="right" height="126"/></a>
 
 <!-- badges: start -->
 
-[![](https://www.r-pkg.org/badges/version/BlueCarbon)](https://cran.r-project.org/package=BlueCarbon)
-![](https://img.shields.io/github/r-package/v/EcologyR/BlueCarbon)
-[![R-CMD-check](https://github.com/EcologyR/BlueCarbon/actions/workflows/R-CMD-check.yaml/badge.svg)](https://github.com/EcologyR/BlueCarbon/actions/workflows/R-CMD-check.yaml)
-[![codecov](https://codecov.io/gh/EcologyR/BlueCarbon/graph/badge.svg)](https://app.codecov.io/gh/EcologyR/BlueCarbon)
-[![](https://img.shields.io/badge/lifecycle-stable-brightgreen.svg)](https://lifecycle.r-lib.org/articles/stages.html#stable)
-[![Project Status: Active - The project has reached a stable, usable
-state and is being actively
-developed.](https://www.repostatus.org/badges/latest/active.svg)](https://www.repostatus.org/#active)
-[![](https://cranlogs.r-pkg.org/badges/grand-total/BlueCarbon)](https://cran.r-project.org/package=BlueCarbon)
+[![](https://www.r-pkg.org/badges/version/BlueCarbon)](https://cran.r-project.org/package=BlueCarbon) ![](https://img.shields.io/github/r-package/v/EcologyR/BlueCarbon) [![R-CMD-check](https://github.com/EcologyR/BlueCarbon/actions/workflows/R-CMD-check.yaml/badge.svg)](https://github.com/EcologyR/BlueCarbon/actions/workflows/R-CMD-check.yaml) [![codecov](https://codecov.io/gh/EcologyR/BlueCarbon/graph/badge.svg)](https://app.codecov.io/gh/EcologyR/BlueCarbon) [![](https://img.shields.io/badge/lifecycle-stable-brightgreen.svg)](https://lifecycle.r-lib.org/articles/stages.html#stable) [![Project Status: Active - The project has reached a stable, usable state and is being actively developed.](https://www.repostatus.org/badges/latest/active.svg)](https://www.repostatus.org/#active) [![](https://cranlogs.r-pkg.org/badges/grand-total/BlueCarbon)](https://cran.r-project.org/package=BlueCarbon)
 
 <!-- badges: end -->
 
-The goal of BlueCarbon is to facilitate the estimation of organic carbon
-stocks and sequestration rates from soil/sediment cores from blue carbon
-ecosystems. Following the protocols published by the Blue Carbon
-initiative [Howard et
-al. (2014)](https://www.cifor-icraf.org/knowledge/publication/5095/).
+The goal of BlueCarbon is to facilitate the estimation of organic carbon stocks and sequestration rates from soil/sediment cores from blue carbon ecosystems. Following the protocols published by the Blue Carbon initiative [Howard et al. (2014)](https://www.cifor-icraf.org/knowledge/publication/5095/).
 
-It includes seven main
-[functions](https://ecologyr.github.io/BlueCarbon/reference/index.html)
-to (1) estimate core compaction, (2) correct core compaction, (3)
-estimate sample thickness, (4) estimate organic carbon content from
-organic matter content, (5) estimate organic carbon stocks and (6)
-sequestration rates, and (7) visualize the error in stock extrapolation.
[…]
