### 2025-10-16 #8333 @goldingn
Hi @costavale you should only need to make that change in [the paper.md file](https://github.com/EcologyR/BlueCarbon/blob/master/manuscript/paper.md).

Then (@NPJuncal and team) can create the release and archive (as per the steps here: https://github.com/openjournals/joss-reviews/issues/8333#issuecomment-3359273007), making sure the affiliation is listed in the same way there.
