### 2025-11-03 c7415a4
add r-universe badges
