### 2025-11-03 bff7c4a
delete intermediate md
