+++ README.knit.md
@@ -1,147 +1,0 @@
----
-output: github_document
-editor_options: 
-  markdown: 
-    wrap: sentence
----
-
-<!-- README.md is generated from README.Rmd. Please edit that file -->
-
-
-
-# BlueCarbon <a href="https://ecologyr.github.io/BlueCarbon/"><img src="man/figures/logo.png" alt="BlueCarbon website" align="right" height="126"/></a>
-
-<!-- badges: start -->
-
-[![](https://www.r-pkg.org/badges/version/BlueCarbon)](https://cran.r-project.org/package=BlueCarbon) ![](https://img.shields.io/github/r-package/v/EcologyR/BlueCarbon) [![R-CMD-check](https://github.com/EcologyR/BlueCarbon/actions/workflows/R-CMD-check.yaml/badge.svg)](https://github.com/EcologyR/BlueCarbon/actions/workflows/R-CMD-check.yaml) [![codecov](https://codecov.io/gh/EcologyR/BlueCarbon/graph/badge.svg)](https://app.codecov.io/gh/EcologyR/BlueCarbon) [![](https://img.shields.io/badge/lifecycle-stable-brightgreen.svg)](https://lifecycle.r-lib.org/articles/stages.html#stable) [![Project Status: Active - The project has reached a stable, usable state and is being actively developed.](https://www.repostatus.org/badges/latest/active.svg)](https://www.repostatus.org/#active) [![](https://cranlogs.r-pkg.org/badges/grand-total/BlueCarbon)](https://cran.r-project.org/package=BlueCarbon)
-
-<!-- badges: end -->
-
-The goal of BlueCarbon is to facilitate the estimation of organic carbon stocks and sequestration rates from soil/sediment cores from blue carbon ecosystems.
-Following the protocols published by the Blue Carbon initiative [Howard et al. (2014)](https://www.cifor-icraf.org/knowledge/publication/5095/).
-
-It includes seven main [functions](https://ecologyr.github.io/BlueCarbon/reference/index.html) to (1) estimate core compaction, (2) correct core compaction, (3) estimate sample thickness, (4) estimate organic carbon content from organic matter content, (5) estimate organic carbon stocks and (6) sequestration rates, and (7) visualize the error in stock extrapolation.
-
-![Blue Carbon package workflow](images/BC_workflow-01.png){width="527"}
-
-#### ***estimate_compaction*** **- Estimate Core Compaction**
-
-Sampling soil cores by manual percussion often results in the compaction of the material retrieved.
-This function (`estimate_compaction()`) estimates the percentage of compaction using measurements taken before and after inserting the corer tube (Fig. 2): the length of the corer tube (sampler_length), distance between the surface of the soil and the top of the tube in the outside (external_distance) and distance between the surface of the soil and the top of the tube in the inside of the tube (internal_distance).
-
-![Soil compaction from field sampling](images/compaction.png){width="429"}
-
-#### ***decompact*** **- Calculate sediment properties after decompaction**
-
-Core compaction derived from field extraction can be mathematically corrected to estimate the original depth of the samples.
-This function (`decompact()`) applies a linear correction (assuming uniform compaction of the core material) to adjust the sample depth accurately.
-If dry bulk density was estimated before compaction correction, it can be provided as a variable and the function will correct it accordingly.
-
-#### ***estimate_oc*** **- Organic carbon % estimation from organic matter %**
-
-There is a linear correlation between organic carbon and organic matter content.
-This correlation can vary across ecosystems and sampling sites.
-This function (`estimate_oc()`) fits a linear regression model between organic matter and organic carbon content of the samples and predicts organic carbon values for samples where the latter information is missing.
[…]
