### 2025-11-03 14b70f0
update images in readme and tutorial #61 #51
