+++ DESCRIPTION
@@ -18,7 +18,7 @@ Description: Tools to estimate soil organic carbon stocks and sequestration rate
 License: GPL (>= 3)
 Encoding: UTF-8
 Roxygen: list(markdown = TRUE)
-RoxygenNote: 7.3.2
+RoxygenNote: 7.3.3
 URL: https://github.com/EcologyR/BlueCarbon, https://ecologyr.github.io/BlueCarbon/
 BugReports: https://github.com/EcologyR/BlueCarbon/issues
 Depends: 
