### 2025-11-03 cfd25fc
use proper vignette instead of web article #50
