+++ pkgdown.yml
@@ -2,8 +2,8 @@ pandoc: 3.1.11
 pkgdown: 2.1.3
 pkgdown_sha: ~
 articles:
-  articles/Tutorial: Tutorial.html
-last_built: 2025-11-03T12:38Z
+  Tutorial: Tutorial.html
+last_built: 2025-11-03T14:11Z
 urls:
   reference: https://ecologyr.github.io/BlueCarbon/reference
   article: https://ecologyr.github.io/BlueCarbon/articles
