### 2025-11-03 2d34db5
Update CITATION.cff
