### 2025-11-03 928ce6c
update news
