+++ NEWS.md
@@ -1,3 +1,7 @@
+# BlueCarbon 0.1.1
+
+* Fully revised documentation and tutorial, following suggestions from @smwindecker and @Tania-Maxwell.
+
 # BlueCarbon 0.1.0
 
 * Package fully revised and ready for public release.
