### 2025-11-03 EcologyR/BlueCarbon#50 @Pakillo
Web article changed to standard vignette, which will be sent to CRAN on next submission
