### 2025-11-03 b8339a1
update citation
