### 2025-11-03 9758729
remove vignettes from Rbuildignore #50
