### 2025-11-03 11ecc9f
Update CITATION.cff
