+++ CITATION.cff
@@ -58,7 +58,7 @@ preferred-citation:
     orcid: https://orcid.org/0000-0002-7981-1599
   year: '2025'
   url: https://github.com/EcologyR/BlueCarbon
-  notes: R package version 0.1.0, https://EcologyR.github.io/BlueCarbon/
+  notes: R package version 0.1.1, https://EcologyR.github.io/BlueCarbon/
 repository: https://CRAN.R-project.org/package=BlueCarbon
 repository-code: https://github.com/EcologyR/BlueCarbon
 url: https://ecologyr.github.io/BlueCarbon/
