### 2025-11-03 efe9d91
update maintainer's email. Close #66
