+++ DESCRIPTION
@@ -2,7 +2,7 @@ Package: BlueCarbon
 Title: Estimation of Organic Carbon Stocks and Sequestration Rates from Soil Core Data
 Version: 0.1.1
 Authors@R: 
-  c(person("Nerea", "Piñeiro-Juncal", email = "nerea.pineiro.juncal@uvigo.gal", 
+  c(person("Nerea", "Piñeiro-Juncal", email = "np.juncal@gmail.com", 
         role = c("aut", "cre", "cph"),
         comment = c(ORCID = "0000-0003-3767-1812")),
     person("Julen", "Astigarraga", role = "aut", comment = c(ORCID = "0000-0001-9520-3713")),
