### 2025-11-03 627fd65
update cran comments
