+++ manuscript/paper.md
@@ -15,16 +15,16 @@ authors:
     affiliation: 1
   - name: Julen Astigarraga
     orcid: 0000-0001-9520-3713
-    affiliation: 2
+    affiliation: 2,3
   - name: Valentina Costa
     orcid: 0000-0002-1513-0284
-    affiliation: 3
+    affiliation: 4,5,6
   - name: Márcio Martins
     orcid: 0000-0002-6969-2215
-    affiliation: 4
+    affiliation: 7
   - name: Francisco Rodríguez-Sánchez
     orcid: 0000-0002-7981-1599
-    affiliation: 5
+    affiliation: 8
 
 
 
@@ -34,15 +34,20 @@ affiliations:
   - index: 2
     name: Universidad de Alcalá, Grupo de Ecología Forestal y Restauración (FORECO), Departamento de Ciencias de la Vida, Spain
   - index: 3
-    name: Stazione Zoologica Anton Dohrn – CRIMAC, Calabria Marine Centre, Department of Integrative Marine Ecology, Amendolara (CS), Italy
-
+    name: Department of Physical Geography and Ecosystem Science, Lund University, Lund, Sweden
   - index: 4
-    name: Centre of Marine Sciences (CCMAR/CIMAR LA), Campus de Gambelas, Universidade do Algarve, 8005-139 Faro, Portugal
+    name: Stazione Zoologica Anton Dohrn – CRIMAC, Calabria Marine Centre, Department of Integrative Marine Ecology, Amendolara (CS), Italy
   - index: 5
-    name: Departamento de Biología Vegetal y Ecología, Universidad de Sevilla, Spain
+    name: National Institute of Oceanography and Applied Geophysics – OGS, Italy
+  - index: 6
+    name: NBFC, National Biodiversity Future Center, Palermo, Italy
+  - index: 7
+    name: Centro de Ciências do Mar do Algarve (CCMAR/CIMAR LA), Campus de Gambelas, Universidade do Algarve, 8005-139 Faro, Portugal
+  - index: 8
+    name: Departamento de Biología Vegetal y Ecología, Facultad de Biología, Universidad de Sevilla, Spain
 
 
-date: 24 February 2025
+date: 3 November 2025
 year: 2025
 bibliography: paper.bib
 csl: apa.csl
@@ -66,7 +71,7 @@ Nerea Piñeiro-Juncal, University of Vigo, Vigo, Spain. Email: [np.juncal\@gmail
 
 # Statement of Need
 
-Coastal blue carbon ecosystems have earned significant attention for their role as organic carbon sinks. Over the past decade, publications on blue carbon research have grown exponentially [@quevedo_2023]. While soil samples can be collected by different methods, estimation methodologies remain fairly homogeneous, following the protocols published by the Blue Carbon initiative [@Howard_2014]. Despite the increasing use of R among blue carbon researchers, there are no specialized R packages dedicated to these calculations. *`BlueCarbon`* aims to standardize and automate the main estimations for calculating  soil and sediment blue carbon stocks and sequestration rates from raw field and laboratory data.
+Coastal blue carbon ecosystems have earned significant attention for their role as organic carbon sinks. Over the past decade, publications on blue carbon research have grown exponentially [@quevedo_2023]. While soil samples can be collected by different methods, estimation methodologies remain fairly homogeneous, following the protocols published by the Blue Carbon initiative [@Howard_2014]. Despite the increasing use of R among blue carbon researchers, there are no specialized R packages dedicated to these calculations. *`BlueCarbon`* aims to standardize and automate the main estimations for calculating soil and sediment blue carbon stocks and sequestration rates from raw field and laboratory data.
 
 # Design
 
