### 2025-11-03 052b7a9
Update CITATION.cff
