### 2025-11-03 c1e0fc6
Deploying to gh-pages from @ EcologyR/BlueCarbon@6c97d7cd32ea221e8fa4bb7933f3503d2166f6a6 🚀
