+++ reference/index.md
@@ -1,0 +1,23 @@
+# Package index
+
+## All functions
+
+- [`bluecarbon_data`](https://ecologyr.github.io/BlueCarbon/reference/bluecarbon_data.md)
+  : Dataframe containing geochemical data from blue carbon soil cores
+- [`core_comp`](https://ecologyr.github.io/BlueCarbon/reference/core_comp.md)
+  : Dataframe containing field measurements of soil cores to estimate
+  compaction caused by core collection.
+- [`decompact()`](https://ecologyr.github.io/BlueCarbon/reference/decompact.md)
+  : Calculate sediment properties after decompaction
+- [`estimate_compaction()`](https://ecologyr.github.io/BlueCarbon/reference/estimate_compaction.md)
+  : Estimate Core Compaction
+- [`estimate_h()`](https://ecologyr.github.io/BlueCarbon/reference/estimate_h.md)
+  : Estimation of the thickness of the sample
+- [`estimate_oc()`](https://ecologyr.github.io/BlueCarbon/reference/estimate_oc.md)
+  : Estimate organic carbon content
+- [`estimate_oc_stock()`](https://ecologyr.github.io/BlueCarbon/reference/estimate_oc_stock.md)
+  : Organic Carbon Stock estimation
+- [`estimate_seq_rate()`](https://ecologyr.github.io/BlueCarbon/reference/estimate_seq_rate.md)
+  : Estimate the average organic carbon sequestration rate
+- [`test_extrapolation()`](https://ecologyr.github.io/BlueCarbon/reference/test_extrapolation.md)
+  : Test differences between observed and extrapolated stocks
