+++ reference/BlueCarbon-package.md
@@ -1,0 +1,34 @@
+# BlueCarbon: Estimation of Organic Carbon Stocks and Sequestration Rates from Soil Core Data
+
+Tools to estimate soil organic carbon stocks and sequestration rates in
+blue carbon ecosystems. 'BlueCarbon' contains functions to estimate and
+correct for core compaction, estimate sample thickness, estimate organic
+carbon content from organic matter content, estimate organic carbon
+stocks and sequestration rates, and visualize the error of carbon stock
+extrapolation.
+
+## See also
+
+Useful links:
+
+- <https://github.com/EcologyR/BlueCarbon>
+
+- <https://ecologyr.github.io/BlueCarbon/>
+
+- Report bugs at <https://github.com/EcologyR/BlueCarbon/issues>
+
+## Author
+
+**Maintainer**: Nerea Piñeiro-Juncal <np.juncal@gmail.com>
+([ORCID](https://orcid.org/0000-0003-3767-1812)) \[copyright holder\]
+
+Authors:
+
+- Julen Astigarraga ([ORCID](https://orcid.org/0000-0001-9520-3713))
+
+- Valentina Costa ([ORCID](https://orcid.org/0000-0002-1513-0284))
+
+- Marcio Martins ([ORCID](https://orcid.org/0000-0002-6969-2215))
+
+- Francisco Rodriguez-Sanchez
+  ([ORCID](https://orcid.org/0000-0002-7981-1599))
