+++ articles/index.md
@@ -1,0 +1,6 @@
+# Articles
+
+### All vignettes
+
+- [BlueCarbon
+  Tutorial](https://ecologyr.github.io/BlueCarbon/articles/Tutorial.md):

[…]
