### 2025-11-20 #8333 @goldingn
So close!

Sorry for the delay. I've read through the manuscript (the previous render) and checked all the things I need to check, and it all looks great. It's just this error that we need to resolve, then I can check rendering again and recommend acceptance.

It looks like a simple one: when setting multiple affiliations they need to be quoted, so: `"2,3"` not `2,3`.
The error appears to be on both [this line](https://github.com/EcologyR/BlueCarbon/blob/4ede78f11f4481931f9e70a677f63c74c9e13f4c/manuscript/paper.md?plain=1#L18) and on [this line](https://github.com/EcologyR/BlueCarbon/blob/4ede78f11f4481931f9e70a677f63c74c9e13f4c/manuscript/paper.md?plain=1#L21)

[Here](https://github.com/zhd52/AOUSDOHtools/blob/a7f596f9143cd7791047040b23c68a58329e2098/paper.md?plain=1#L21) is an example from another submission that renders fine.
