### 2025-11-20 9746003
correct afiliation rendering
