### 2025-11-21 25ec51d
fix Howard citation
