+++ manuscript/paper.bib
@@ -18,7 +18,7 @@
 @book{Howard_2014,
   	url = {https://www.thebluecarboninitiative.org/manual},
   	Author = {Howard, J. and Hoyt, S. and Isensee, K. and Pidgeon, E. and Telszewski, M.},
-  	Booktitle = {Coastal Blue Carbon: Methods for assessing carbon stocks and emissions factors in mangroves, tidal salt marshes, and seagrass meadows.},
+  	title = {Coastal Blue Carbon: Methods for assessing carbon stocks and emissions factors in mangroves, tidal salt marshes, and seagrass meadows.},
   	Publisher = {Conservation International, Intergovernmental Oceanographic Commission of UNESCO, International Union for Conservation of Nature},
   	Year = 2014
 }
