+++ manuscript/paper.tex
@@ -478,7 +478,9 @@ doi:\href{https://doi.org/10.1038/ngeo1477}{10.1038/ngeo1477}
 
 \bibitem[\citeproctext]{ref-Howard_2014}
 Howard, J., Hoyt, S., Isensee, K., Pidgeon, E., \& Telszewski, M.
-(2014). Conservation International, Intergovernmental Oceanographic
+(2014). \emph{Coastal blue carbon: Methods for assessing carbon stocks
+and emissions factors in mangroves, tidal salt marshes, and seagrass
+meadows.} Conservation International, Intergovernmental Oceanographic
 Commission of UNESCO, International Union for Conservation of Nature.
 Retrieved from \url{https://www.thebluecarboninitiative.org/manual}
 
