### 2025-11-26 fb5023e
update citation
