### 2025-11-26 7adad00
citation
