### 2025-11-26 0c57b4c
update citation info
