+++ NEWS.md
@@ -1,3 +1,7 @@
+# BlueCarbon 0.1.2
+
+* Update citation info, after publication in the Journal of Open Source Software.
+
 # BlueCarbon 0.1.1
 
 * Fully revised documentation and tutorial, following suggestions from @smwindecker and @Tania-Maxwell.
