+++ cran-comments.md
@@ -1,13 +1,5 @@
 ## R CMD check results
 
-0 errors | 0 warnings | 1 note
+0 errors | 0 warnings | 0 note
 
-❯ checking CRAN incoming feasibility ... [8s/27s] NOTE
-  Maintainer: ‘Nerea Piñeiro-Juncal <np.juncal@gmail.com>’
-  
-  New maintainer:
-    Nerea Piñeiro-Juncal <np.juncal@gmail.com>
-  Old maintainer(s):
-    Nerea Piñeiro-Juncal <nerea.pineiro.juncal@uvigo.gal>
-
-Please note the maintainer's email has changed, since the 'uvigo.gal' email address will likely be closed soon.
+Updated citation information.
