+++ README.Rmd
@@ -116,15 +116,26 @@ remotes::install_github("EcologyR/BlueCarbon")
 
 If using this package, please cite it:
 
-To cite package 'BlueCarbon' in publications use:
-
-Piñeiro-Juncal et al., (2025).
-BlueCarbon R package: Estimation of Organic Carbon Stocks and Sequestration Rates From Soil Core Data.
-Journal of Open Source Software, 10(115), 8333, <https://doi.org/10.21105/joss.08333>
+Piñeiro-Juncal et al., (2025). BlueCarbon R package: Estimation of Organic Carbon Stocks and Sequestration Rates From Soil Core Data. Journal of Open Source Software, 10(115), 8333, <https://doi.org/10.21105/joss.08333>
 
 A BibTeX entry for LaTeX users is
 
-@article{Piñeiro-Juncal2025, doi = {10.21105/joss.08333}, url = {<https://doi.org/10.21105/joss.08333>}, year = {2025}, publisher = {The Open Journal}, volume = {10}, number = {115}, pages = {8333}, author = {Piñeiro-Juncal, Nerea and Astigarraga, Julen and Costa, Valentina and Martins, Márcio and Rodríguez-Sánchez, Francisco}, title = {BlueCarbon R package: Estimation of Organic Carbon Stocks and Sequestration Rates From Soil Core Data}, journal = {Journal of Open Source Software} }
+```{verbatim}
+@article{Piñeiro-Juncal2025,
+  doi = {10.21105/joss.08333},
+  url = {https://doi.org/10.21105/joss.08333},
+  year = {2025},
+  publisher = {The Open Journal},
+  volume = {10},
+  number = {115},
+  pages = {8333},
+  author = {Piñeiro-Juncal, Nerea and Astigarraga, Julen and Costa, Valentina and Martins, Márcio and Rodríguez-Sánchez, Francisco},
+  title = {BlueCarbon R package: Estimation of Organic Carbon Stocks and Sequestration Rates From Soil Core Data},
+  journal = {Journal of Open Source Software}
+}
+```
+
+
 
 ## Code of Conduct
 
