+++ .github/workflows/joss.yaml
@@ -1,28 +1,0 @@
-name: Draft PDF
-on:
-  workflow_dispatch:
-  push:
-    paths:
-    - 'manuscript/**'
-
-jobs:
-  paper:
-    runs-on: ubuntu-latest
-    name: Paper Draft
-    steps:
-      - name: Checkout
-        uses: actions/checkout@v6
-      - name: Build draft PDF
-        uses: openjournals/openjournals-draft-action@master
-        with:
-          journal: joss
-          # This should be the path to the paper within your repo.
-          paper-path: manuscript/paper.md
-      - name: Upload
-        uses: actions/upload-artifact@v5
-        with:
-          name: paper
-          # This is the output path where Pandoc will write the compiled
-          # PDF. Note, this should be the same directory as the input
-          # paper.md
-          path: manuscript/paper.pdf

[…]
